refactor(core): rename model_profiles to llm_model_profiles
Rename module for consistency with other LLM-related modules:
- llm_drivers.rs - LLM driver abstractions
- llm_models.rs - LLM domain entities
- llm_model_profiles.rs - Model capability profiles
