+++ crates/everruns-core/src/llm_model_profiles.rs
